# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Liste toutes les locations",
                "description": "Display a listing of the resource.",
                "operationId": "4a95f25c1e13530115fd3fd6961b92de",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Créer une nouvelle location",
                "description": "Store a newly created resource in storage.",
                "operationId": "2e940d1dad83fe7c7fb5a308a7f6d84a",
                "requestBody": {
                    "description": "Données de la location",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Location"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location créée avec succès"
                    }
                }
            }
        },
        "/api/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Récupérer une location par ID",
                "description": "Display the specified resource.",
                "operationId": "5ac0324577506ee36225fdb6a881506c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Location non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Mettre à jour une location",
                "description": "Update the specified resource in storage.",
                "operationId": "42f75cf2951683141c55565f6c6e10ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données de la location à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Location"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location mise à jour avec succès"
                    },
                    "404": {
                        "description": "Location non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Supprimer une location",
                "description": "Remove the specified resource from storage.",
                "operationId": "32274c410d9a158665aabbb58f7830ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Location supprimée avec succès"
                    },
                    "404": {
                        "description": "Location non trouvée"
                    }
                }
            }
        },
        "/api/matches": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Liste toutes les correspondances",
                "operationId": "a3b0ee8d77fd59c8b81034a43142f1a3",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Matches"
                ],
                "summary": "Créer une nouvelle correspondance",
                "operationId": "06c20c0bdb68eb85dfc41da491d455a6",
                "requestBody": {
                    "description": "Données de la correspondance",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "App\\Models\\Matche"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Correspondance créée avec succès"
                    }
                }
            }
        },
        "/api/matches/{id}": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Récupérer une correspondance par ID",
                "operationId": "94139b056e06bc8205dcbac924dc2a24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la correspondance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Correspondance non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "Mettre à jour une correspondance",
                "operationId": "a789c418e6ae8b469fd4e2b58ec56b6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la correspondance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données de la correspondance à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "App\\Models\\Matche"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Correspondance mise à jour avec succès"
                    },
                    "404": {
                        "description": "Correspondance non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Matches"
                ],
                "summary": "Supprimer une correspondance",
                "operationId": "de602074ac0ab8130bad58ee90e61a95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la correspondance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Correspondance supprimée avec succès"
                    },
                    "404": {
                        "description": "Correspondance non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Location": {
                "properties": {
                    "city": {
                        "type": "string",
                        "example": ""
                    },
                    "country": {
                        "type": "string",
                        "example": ""
                    },
                    "latitude": {
                        "type": "string",
                        "example": ""
                    },
                    "longitude": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}